% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ.R
\name{categ}
\alias{categ}
\alias{r_categ}
\alias{r_categ_mat}
\title{Categorical Distribution}
\usage{
r_categ(n, p, log_p = FALSE)

r_categ_mat(P, log_p = FALSE)
}
\arguments{
\item{n}{Number of desired draws}

\item{p}{A \eqn{k}-dimensional vector of probabilities}

\item{log_p}{If \code{TRUE}, interpret given probabilities as having been
supplied on the log-scale. Otherwise, interpret them on the original scale.}

\item{P}{An \eqn{n \times k} matrix of probabilities, where \eqn{k} is
the number of categories and \eqn{n} is the desired number of independent
draws.}
}
\value{
a vector of category zero-based indices whose elements are in
\eqn{0, \ldots, k-1}.
}
\description{
Sample from a matrix of categorical probabilities.
}
\details{
It might be more useful to have a version of this function that takes
probabilities on the log-scale. We could consider using the Gumbel trick
for that.
}
