% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{\%notin\%}
\alias{\%notin\%}
\title{"Not In" Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{Vector or NULL: the values to be matched. Long vectors are
supported.}

\item{table}{Vector or NULL: the values to be matched against. Long vectors
are not supported.}
}
\value{
The negation of the \code{\%in\%} operator.
}
\description{
"Not In" Operator
}
