% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holm.R
\name{holm}
\alias{holm}
\title{Holm's Procedure for Multiple Comparisons}
\usage{
holm(p, alpha)
}
\arguments{
\item{p}{vector of \eqn{m} p-values.}

\item{alpha}{Family-wise error rate.}
}
\value{
Vector of indices \eqn{\mathcal{I}} such that
\eqn{H_{0i}: i \in \mathcal{I}} can be rejected. Result is empty if none can
be rejected.
}
\description{
Holm's Procedure for Multiple Comparisons
}
\examples{
p = c(0.101, 0.05, 0.03)
holm(p, 0.15)
holm(p, 0.10)
holm(p, 0.05)

}
\references{
Holm, S. (1979). A Simple Sequentially Rejective Multiple Test Procedure.
Scandinavian Journal of Statistics, 6(2), 65-70.
}
