% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-temp-pkg.R
\name{import_temp_pkg}
\alias{import_temp_pkg}
\title{Create and Import a Temporary Package}
\usage{
import_temp_pkg(src_files, pkg_name, unlink_on_exit = TRUE)
}
\arguments{
\item{src_files}{A vector of filenames with source code for the temp package}

\item{pkg_name}{A name for the temp package}

\item{unlink_on_exit}{Remove the temporary package when R exits}
}
\value{
Name of the temporary directory with the package
}
\description{
Create a temporary package from source files so that functions can be
accessed from a namespace. This is based on Josh O'Brien's answer from
\url{https://stackoverflow.com/questions/15789036/namespaces-without-packages}
}
\examples{
# Create a couple of example source files
ff1 = tempfile(pattern = "bar", fileext = ".R")
ff2 = tempfile(pattern = "baz", fileext = ".R")
cat("bar = function() { print('Hello world') }", file = ff1)
cat("baz = function() { print('Goodbye world.') }", file = ff2)
import_temp_pkg(src_files = c(ff1, ff2), pkg_name = "foo")
unlink(c(ff1,ff2))

}
