% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{InverseGamma}
\alias{InverseGamma}
\alias{r_invgamma}
\alias{d_invgamma}
\title{Inverse Gamma distribution}
\usage{
r_invgamma(n, a, b)

d_invgamma(x, a, b, log = FALSE)
}
\arguments{
\item{n}{Number of observations.}

\item{a}{Shape parameter.}

\item{b}{Rate parameter.}

\item{x}{Vector of quantiles.}

\item{log}{If \code{TRUE}, return densities and probabilities on the log-scale.}
}
\value{
\code{dinvgamma} gives the density, \code{rinvgamma} generates random
deviates.
}
\description{
Inverse Gamma distribution
}
