% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-sum-exp.R
\name{Log-Sum-Exp}
\alias{Log-Sum-Exp}
\alias{log_sum_exp}
\alias{log_add2_exp}
\alias{log_sub2_exp}
\title{Log-Sum-Exp}
\usage{
log_sum_exp(x)

log_add2_exp(x, y)

log_sub2_exp(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\description{
Compute \code{log(sum(exp(x)))} but in a more stable way.
}
\details{
Computed using the method described by user Ben in StackExchange
thread \url{https://stats.stackexchange.com/questions/381936/vectorised-computation-of-logsumexp}.
A faster C version (requiring a dependency and possibly compilation) is
provided in \link[matrixStats]{logSumExp}.
}
\examples{
pi = 1:6 / sum(1:6)
x = log(2*pi)
log(sum(exp(x)))
log_sum_exp(x)

}
