% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh-rw-txform.R
\name{mh_rw_txform}
\alias{mh_rw_txform}
\alias{print.mh_rw_txform}
\alias{parnames.mh_rw_txform}
\alias{parnames<-.mh_rw_txform}
\title{Random Walk Metropolis-Hastings Transformed Parameters}
\usage{
\method{print}{mh_rw_txform}(x, ...)

\method{parnames}{mh_rw_txform}(object, ...)

\method{parnames}{mh_rw_txform}(object) <- value
}
\arguments{
\item{x}{Object to print}

\item{...}{Additional arguments}

\item{object}{Object to operate on}

\item{value}{Vector of new parameter names}
}
\description{
These methods operate on results of \code{txform} from \code{mh_rw}.
}
\examples{
# TBD
}
