% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-optim-ci.R
\name{mle_optim_ci}
\alias{mle_optim_ci}
\alias{print.mle_optim_ci}
\title{MLE Optimization Confidence Intervals}
\usage{
\method{print}{mle_optim_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mle_optim_ci}}

\item{...}{Additional arguments}
}
\description{
MLE Optimization Confidence Intervals
}
