% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-txform.R
\name{mle_optim_txform}
\alias{mle_optim_txform}
\alias{print.mle_optim_txform}
\alias{coef.mle_optim_txform}
\alias{vcov.mle_optim_txform}
\alias{parnames.mle_optim_txform}
\alias{parnames<-.mle_optim_txform}
\title{MLE Optimization Transformed Parameters}
\usage{
\method{print}{mle_optim_txform}(x, ...)

\method{coef}{mle_optim_txform}(object, ...)

\method{vcov}{mle_optim_txform}(object, ...)

\method{parnames}{mle_optim_txform}(object, ...)

\method{parnames}{mle_optim_txform}(object) <- value
}
\arguments{
\item{x}{Object to operate on}

\item{...}{Additional arguments}

\item{object}{Object to pprint}

\item{value}{Vector of new parameter names}
}
\description{
These methods operate on results of \code{txform} from \code{mle_optim}.
}
