% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R
\name{mlogit}
\alias{mlogit}
\alias{inv_mlogit}
\title{Multinomial logit transformation}
\usage{
mlogit(p, ref = 1)

inv_mlogit(x, ref = 1)
}
\arguments{
\item{p}{A point in the probability simplex \eqn{S^J}.}

\item{ref}{Index of the reference category. Default is \code{1}, the first
category.}

\item{x}{A point in Euclidean space \eqn{R^{J-1}}.}
}
\description{
The multinomial logit transformation and its inverse.
}
\details{
\itemize{
\item{mlogit}{Transforms from probability simplex \eqn{S^J} to \eqn{R^{J-1}}.}
\item{inv_mlogit}{Transforms from \eqn{R^{J-1}} to probability simplex \eqn{S^J}.}
}
}
\examples{
p = 1:5 / sum(1:5)
x = mlogit(p)
p_new = inv_mlogit(x)

}
