% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ.R
\name{mult2cat}
\alias{mult2cat}
\title{mult2cat}
\usage{
mult2cat(x)
}
\arguments{
\item{x}{A \eqn{k \times n} matrix of counts}
}
\value{
An \eqn{n} dimensional vector of indices representing categories of
individual trials.
}
\description{
Ungroup a \eqn{k \times n} matrix of multinomial observations to individual
categories, where \eqn{k} is the number of categories abd \eqn{n} is the
number of observations.
}
