% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{na}
\alias{na}
\title{na}
\usage{
na(length, type = c("real", "integer", "complex", "character"))
}
\arguments{
\item{length}{A non-negative integer specifying the desired length.}

\item{type}{A string matching one of: \code{character}, \code{complex},
\code{integer}, or \code{real}}
}
\description{
Create a vector of NA values.
}
