% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{parnames}
\alias{parnames}
\alias{parnames<-}
\title{Parameter names}
\usage{
parnames(object, ...)

parnames(object) <- value
}
\arguments{
\item{object}{An object.}

\item{...}{Additional parameters.}

\item{value}{a character vector.}
}
\description{
Get or set parameter names.
}
