% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{polar}
\alias{polar}
\alias{inv_polar}
\title{Polar transformation}
\usage{
polar(z)

inv_polar(x)
}
\arguments{
\item{z}{A point in the unit sphere \eqn{\mathbb{S}^d = \{ \bm{z} \in \mathbb{R}^d: \bm{z}^\top \bm{z} = 1 \}}.}

\item{x}{A point in Euclidean space \eqn{\mathbb{R}^{d-1}}.}
}
\description{
The polar transformation and its inverse.
}
\details{
\describe{
\item{polar}{Transforms from  \eqn{\mathbb{S}^d} to \eqn{\mathbb{R}^{d-1}}.}
\item{inv_polar}{Transforms from \eqn{\mathbb{R}^{d-1}} to \eqn{\mathbb{S}^d}.}
}
}
\examples{
p = 1:5 / sum(1:5)
x = mlogit(p)
p_new = inv_mlogit(x)

}
