% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{Print}
\alias{Print}
\alias{printf}
\alias{logger}
\alias{fprintf}
\alias{my_numerical_format}
\title{Printing}
\usage{
printf(fmt, ...)

logger(fmt, ...)

fprintf(file, fmt, ...)

my_numerical_format(x, tol = 1e-04, fmt_small = "\%0.3E", fmt_large = "\%0.4f")
}
\arguments{
\item{fmt}{Formatted string which can be processed by \code{sprintf}}

\item{...}{Additional arguments}

\item{file}{A connection, or a character string naming the file to print to}

\item{x}{A numeric vector}

\item{tol}{A positive tolerance}

\item{fmt_small}{An \code{sprintf} format to display small-magnitude numbers}

\item{fmt_large}{An \code{sprintf} format to display large-magnitude numbers}
}
\description{
Printing
}
\details{
For \code{my_numerical_format}n numbers in \code{x} with smaller magnitude
than \code{tol} are formatted using \code{fmt_small}; others are formatted
using \code{fmt_large}.
}
