% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-interface.R
\name{print_interface}
\alias{print_interface}
\title{print_function}
\usage{
print_interface(x, ...)
}
\arguments{
\item{x}{A function object}

\item{...}{Additional options passed to \link[stringr:str_wrap]{stringr::str_wrap}}
}
\value{
Nothing is returned
}
\description{
Print the interface of a function in a readable way. Excludes the body of
the function.
}
\details{
If \code{...} argument is empty, pass \code{exdent = 4} to \link[stringr:str_wrap]{stringr::str_wrap}.
Otherwise, use whatever is given in \code{...}.
}
