% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_categ_mat}
\alias{r_categ_mat}
\title{r_categ_mat}
\usage{
r_categ_mat(prob)
}
\arguments{
\item{prob}{An \eqn{n \times k} matrix of probabilities, where \eqn{k} is
the number of categories and \eqn{n} is the desired number of indepedent
draws.}
}
\value{
a vector of category zero-based indices whose elements are in
\eqn{0, \ldots, k-1}.
}
\description{
Sample from a matrix of categorical probabilities.
}
\details{
It might be more useful to have a version of this function that takes
probabilities on the log-scale. We could consider using the Gumbel trick
for that.
}
