% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{r_truncated}
\alias{r_truncated}
\title{r_truncated}
\usage{
r_truncated(n, lo, hi, pf, qf, ...)
}
\arguments{
\item{n}{}

\item{lo}{}

\item{hi}{}

\item{pf}{}

\item{qf}{}

\item{...}{}
}
\description{
Generate draws from a truncated univariate distribution based on its CDF and
quantile function. This is basically copied from
\link[LearnBayes]{rtruncated}.
}
