% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{raim-package}
\alias{raim}
\alias{raim-package}
\title{raim: Andrew's Package of Miscellaneous Functions}
\description{
A package of miscellaneous functions that the author uses.
}
\author{
\strong{Maintainer}: Andrew Raim \email{andrew.raim@gmail.com}

}
