% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqres.R
\name{rqres}
\alias{rqres}
\title{Randomized Quantile Residuals}
\usage{
rqres(y, p_dist, eps = 1e-06)
}
\arguments{
\item{y}{Vector of univariate observations}

\item{p_dist}{A cumulative distribution function. Typically the CDF of a
fitted model with with parameters evaluated at  estimates.}

\item{eps}{A small positive number needed for discrete distributions.
Set to zero to avoid using random jitter.}
}
\value{
Vector of residuals
}
\description{
Compute randomized quantile residuals (Dunn & Smyth, 1996).
}
\examples{
mu_true = 10
y = rpois(200, mu_true)
mu_hat = mean(y)
p_dist = function(x) { ppois(y, mu_hat) }
res = rqres(y, p_dist)

}
\references{
Peter K. Dunn & Gordon K. Smyth (1996) Randomized Quantile Residuals,
Journal of Computational and Graphical Statistics, 5:3, 236-244,
DOI: 10.1080/10618600.1996.10474708.
}
