% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_along_dim}
\alias{seq_along_dim}
\title{Sequence}
\usage{
seq_along_dim(x, d)
}
\arguments{
\item{x}{A matrix, array, or other object which \eqn{dim} can operate on.}

\item{d}{A dimension of \eqn{x}}
}
\description{
Sequence
}
