% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift-scale.R
\name{shift_scale}
\alias{shift_scale}
\alias{inv_shift_scale}
\title{Shift/scale transformation}
\usage{
shift_scale(x, a, b)

inv_shift_scale(u, a, b)
}
\arguments{
\item{x}{A point in \eqn{[a,b]}.}

\item{a}{A number.}

\item{b}{A number.}

\item{u}{A point in \eqn{[0,1]}.}
}
\description{
Functions to Transformations between the unit interval \eqn{[0,1]} and
the interval \eqn{[a,b]} with \eqn{a \leq b}.
}
\examples{
u = seq(0, 1, 0.1)
x = inv_shift_scale(u, 1, 10)
u_new = shift_scale(x, 1, 10)

}
