% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{txform}
\alias{txform}
\title{Transform Estimates}
\usage{
txform(object, tx, ...)
}
\arguments{
\item{object}{a result from an estimation method}

\item{tx}{A function of the parameters estimated in \code{object}}

\item{...}{additional arguments}
}
\value{
Estimates of \code{tx(phi)}
}
\description{
Transform results from an estimation method
}
\details{
Suppose \code{object} is the result of an estimation method that produces
estimates or MCMC draws of model parameters \eqn{\phi} and \code{tx} is a
function \eqn{t(\phi) = (t_1(\phi), \ldots, t_k(\phi))} of \eqn{\phi}. The
\code{txform} function produces corresponding results for \eqn{t(\phi)},
}
