% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{unitvec}
\alias{unitvec}
\title{Unit Vector}
\usage{
unitvec(j, n)
}
\arguments{
\item{j}{Position of one}

\item{n}{Dimension of vector}
}
\value{
An \eqn{n} dimensional vector with a one in the \eqn{j}th position and zeros
in other positions.
}
\description{
Unit Vector
}
