% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{Dirichlet}
\alias{Dirichlet}
\alias{ddirichlet}
\alias{rdirichlet}
\title{Dirichlet Distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

rdirichlet(n, alpha)
}
\arguments{
\item{x}{A value in the sample space}

\item{alpha}{Vector of eqn{k} rate parameters}

\item{n}{Number of desired draws}
}
\value{
A vector of draws
}
\description{
Functions for the Dirichlet distribution
}
