% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{Gumbel}
\alias{Gumbel}
\alias{rgumbel}
\title{Gumbel Distribution}
\usage{
rgumbel(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{Number of desired draws}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}
}
\value{
A vector of draws
}
\description{
Functions for the Gumbel distribution
}
