% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_rw.R
\name{mh_rw}
\alias{mh_rw}
\title{Random Walk Metropolis-Hastings with Gaussian Proposal}
\arguments{
\item{proposal_var}{}

\item{proposal_scale}{}

\item{R}{}

\item{grp}{}

\item{burn}{}

\item{thin}{}

\item{report_period}{}
}
\description{
This code was originally adapted from \code{rwmetrop} function in the
\code{LearnBayes} package
}
