% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh-rw.R
\name{mh_rw}
\alias{mh_rw}
\alias{mh_rw_control}
\alias{summary.mh_rw}
\alias{print.mh_rw}
\alias{parnames.mh_rw}
\alias{parnames<-.mh_rw}
\alias{txform.mh_rw}
\title{Random Walk Metropolis-Hastings with Gaussian Proposal}
\usage{
mh_rw_control(
  proposal_scale = 1,
  R = 1000,
  grp = NULL,
  burn = 0,
  thin = 1,
  report_period = R + 1
)

mh_rw(init, log_post, proposal_var, control = mh_rw_control())

\method{summary}{mh_rw}(object, level = 0.05, ...)

\method{print}{mh_rw}(x, ...)

\method{parnames}{mh_rw}(object, ...)

\method{parnames}{mh_rw}(object) <- value

\method{txform}{mh_rw}(object, tx, ...)
}
\arguments{
\item{proposal_scale}{A scalar multiple applied to \code{proposal_var}}

\item{R}{Length of requested chain}

\item{grp}{Vector of integers specifying grouping of elements of \eqn{\phi}.
See details.}

\item{burn}{Number of initial draws to burn / discard}

\item{thin}{Thinning factor; the period for saving draws}

\item{report_period}{Period for reporting MCMC progress}

\item{init}{Initial value of \eqn{\phi} for the sampler}

\item{log_post}{Log-posterior function of \eqn{\phi} with target distribution}

\item{proposal_var}{Variance matrix of proposal distribution}

\item{control}{An object created via \code{mh_rw_control}}

\item{object}{Result of \code{mh_rw} or \code{txform}}

\item{...}{Additional arguments}

\item{x}{Object to print}

\item{tx}{A transformation of the parameter \eqn{\phi}}
}
\value{
A list with the following elements
\itemize{
\item{par}{Matrix of draws where the \eqn{q} columns correspond to the
parameters of \code{log_post}}
\item{accept}{Vector of acceptance rates for the groups defined in in
\code{grp}}
}
}
\description{
This code was originally adapted from \code{rwmetrop} function in the
\code{LearnBayes} package.
}
\details{
TBD: describe \code{grp}, proposal, \code{log_post}.
}
\examples{
# TBD
}
