% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_optim}
\alias{mle_optim}
\alias{mle_optim_control}
\alias{txform.mle_result}
\alias{print.txform_mle_result}
\alias{coef.mle_transform_result}
\alias{vcov.mle_transform_result}
\alias{confint.mle_result}
\alias{print.mle_result_ci}
\alias{print.mle_result}
\alias{summary.mle_result}
\alias{coef.mle_result}
\alias{logLik.mle_result}
\alias{AIC.mle_result}
\alias{BIC.mle_result}
\alias{vcov.mle_result}
\title{MLE Optimization}
\usage{
mle_optim_control(optim_method = "L-BFGS-B", optim_control = list())

mle_optim(
  init,
  loglik,
  n = NA,
  df = Inf,
  fixed = integer(0),
  control = mle_optim_control()
)

\method{txform}{mle_result}(object, tx, jacobian = NULL, ...)

\method{print}{txform_mle_result}(x, ...)

\method{coef}{mle_transform_result}(object, ...)

\method{vcov}{mle_transform_result}(object, ...)

\method{confint}{mle_result}(object, parm, level = 0.95, ...)

\method{print}{mle_result_ci}(x, ...)

\method{print}{mle_result}(x, ...)

\method{summary}{mle_result}(object, ...)

\method{coef}{mle_result}(object, ...)

\method{logLik}{mle_result}(object, ...)

\method{AIC}{mle_result}(object, ..., k = 2)

\method{BIC}{mle_result}(object, ...)

\method{vcov}{mle_result}(object, ...)
}
\arguments{
\item{optim_method}{A \code{method} argument to pass to \code{optim}}

\item{optim_control}{A \code{control} argument to pass to \code{optim}}

\item{init}{Initial value of \eqn{\phi} for the sampler}

\item{loglik}{Log-likelihood function of \eqn{\phi}}

\item{n}{Sample size. Needed for AIC, BIC and some other optional results.
Default value \code{NA} results in \code{NA} values for those optional
results.}

\item{df}{Degrees of freedom to use for tests. Default value \code{Inf}
assumes a large sample Normal distribution.}

\item{fixed}{a vector of integers; default is an empty vector. These are
interpreted as indices into \code{init}; corresponding elements are held
fixed at their \code{init} values during optimization.}

\item{control}{Object created via \code{mle_optim_control} function.}

\item{object}{Result of \code{mle_optim} or \code{txform}}

\item{tx}{A transformation of the parameter \eqn{\phi}}

\item{jacobian}{}

\item{...}{Additional arguments}

\item{x}{}

\item{parm}{}

\item{level}{}

\item{k}{}
}
\description{
MLE Optimization
}
