% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_optim}
\alias{mle_optim}
\alias{mle_optim_control}
\alias{mle_transform}
\alias{print.mle_transform_result}
\alias{coef.mle_transform_result}
\alias{vcov.mle_transform_result}
\alias{confint.mle_result}
\alias{print.mle_result_ci}
\alias{print.mle_result}
\alias{summary.mle_result}
\alias{coef.mle_result}
\alias{logLik.mle_result}
\alias{AIC.mle_result}
\alias{BIC.mle_result}
\alias{vcov.mle_result}
\title{MLE Optimization}
\usage{
mle_optim_control(optim_method = "L-BFGS-B", optim_control = list())

mle_optim(init, loglik, n, df = n, control = mle_optim_control())

mle_transform(object, tx, labels = NULL)

\method{print}{mle_transform_result}(x, ...)

\method{coef}{mle_transform_result}(object, ...)

\method{vcov}{mle_transform_result}(object, ...)

\method{confint}{mle_result}(object, parm, level = 0.95, ...)

\method{print}{mle_result_ci}(x, ...)

\method{print}{mle_result}(x, ...)

\method{summary}{mle_result}(object, ...)

\method{coef}{mle_result}(object, ...)

\method{logLik}{mle_result}(object, ...)

\method{AIC}{mle_result}(object, ..., k = 2)

\method{BIC}{mle_result}(object, ...)

\method{vcov}{mle_result}(object, ...)
}
\arguments{
\item{optim_method}{}

\item{optim_control}{}
}
\description{
MLE Optimization
}
